float to int + article vendu correction

+++ src/main/java/com/example/encheres/bll/UtilisateurService.java
@@ -6,11 +6,11 @@ import com.example.encheres.bo.Utilisateur;
 import com.example.encheres.exception.BusinessException;
 
 public interface UtilisateurService {
-// methodes sans les utilisateurs historisés	
+// methodes sans les utilisateurs historisés
 	void creerUtilisateur ( Utilisateur utilisateur)  throws BusinessException;
 	Utilisateur lectureUtilisateur(int noUtilisateur);
 	void modifierUtilisateur(Utilisateur utilisateur) throws BusinessException;
-	void modifierUtilisateurCredit(int noUtilisateur, float credit);
+	void modifierUtilisateurCredit(int noUtilisateur, int credit);
 	void supprimerUtilisateur(int noUtilisateur)      throws BusinessException;
 	Utilisateur findByPseudo( String pseudo);
 	List<Utilisateur> findAll ();
@@ -18,6 +18,6 @@ public interface UtilisateurService {
 	// methodes avec les utilisateurs historisés
 	List<Utilisateur> findAllHisto();
 	void historiserUtilisateur (int noUtilisateur)    throws BusinessException;
-	
-	
+
+
 }
